Tidy up of Structure
