+++ src/main.tsx
@@ -2,23 +2,23 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import Root from './routes/Root/Root.tsx'; 
-import LoginWrapper from './components/LoginWrapper';
-import App from './components/Dashboard/App';
+import LoginWrapper from './components/Login/LoginWrapper.tsx';
+import App from './components/Dashboard/DashboardApp.tsx';
 
 const router = createBrowserRouter([
-    {
-        element: <Root />,
-        errorElement: <Root />,
-        children: [
+	{
+		element: <Root />,
+		errorElement: <Root />,
+		children: [
             
-            { path: 'login', element: <LoginWrapper /> },
-            { path: 'dashboard/*', element: <App /> }, // Use App and manage sub-routes inside it
-        ],
-    },
+			{ path: 'login', element: <LoginWrapper /> },
+			{ path: 'dashboard/*', element: <App /> }, // Use App and manage sub-routes inside it
+		],
+	},
 ]);
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-    <React.StrictMode>
-        <RouterProvider router={router} />
-    </React.StrictMode>,
+	<React.StrictMode>
+		<RouterProvider router={router} />
+	</React.StrictMode>,
 );
